+++ runtests.py
@@ -12,43 +12,39 @@ SETTINGS_DICT = {
     'DATABASES': ConfigHelper.generate_database_configs(
         unshard={
             'default': {
-                'master': 'mysql://root:root@127.0.0.1/default?sql_mode=STRICT_TRANS_TABLES&charset=utf8',
+                'master': {'url': 'mysql://root:root@127.0.0.1/default?sql_mode=STRICT_TRANS_TABLES&charset=utf8'},
             },
         },
         shard={
             'shard_a': {
-                'shard_options': {
-                    'database_name': 'shard_a',
-                    'logical_count': 2,
-                },
-                'db_options': {
+                'database_name': 'shard_a',
+                'logical_count': 2,
+                'options': {
                     'sql_mode': 'STRICT_TRANS_TABLES',
                     'charset': 'utf8',
                 },
                 'shards': [
                     {
-                        'master': 'mysql://root:root@127.0.0.1/',
+                        'master': {'url': 'mysql://root:root@127.0.0.1/'},
                     },
                     {
-                        'master': 'mysql://root:root@127.0.0.1/',
+                        'master': {'url': 'mysql://root:root@127.0.0.1/'},
                     },
                 ]
             },
             'shard_b': {
-                'shard_options': {
-                    'database_name': 'shard_b',
-                    'logical_count': 2,
-                },
-                'db_options': {
+                'database_name': 'shard_b',
+                'logical_count': 2,
+                'options': {
                     'sql_mode': 'STRICT_TRANS_TABLES',
                     'charset': 'utf8',
                 },
                 'shards': [
                     {
-                        'master': 'mysql://root:root@127.0.0.1/',
+                        'master': {'url': 'mysql://root:root@127.0.0.1/'},
                     },
                     {
-                        'master': 'mysql://root:root@127.0.0.1/',
+                        'master': {'url': 'mysql://root:root@127.0.0.1/'},
                     },
                 ]
             }