+++ shard/config/database_config.py
@@ -1,60 +1,76 @@
+import copy
 import urllib.parse
-from typing import Optional, Tuple, Dict, List
+from typing import Dict, List, Optional, Tuple
 
 from dj_database_url import config
 
 from shard.constants import DATABASE_CONFIG_MASTER, DATABASE_CONFIG_SHARD_GROUP, DATABASE_CONFIG_SHARD_NUMBER
-from shard.exceptions import RequireMasterConfigException
 
 __all__ = ('make_shard_configuration', 'make_replication_configuration', )
 
 
-def make_shard_configuration(shard_group: str, shard_options: Dict, shards: List, db_options: Optional[Dict]=None) -> Dict:
+def make_shard_configuration(
+        shard_group: str, database_name: str, logical_count: int, conn_max_age: int, shards: List, options: Dict
+) -> Dict:
     configuration = {}
-    database_name = shard_options['database_name']
     shard_count = len(shards)
-    logical_count_per_shard = _get_logical_count_per_shard(logical_count=shard_options['logical_count'], shard_count=shard_count)
+    logical_count_per_shard = _get_logical_count_per_shard(logical_count=logical_count, shard_count=shard_count)
 
-    for index, shard_config in enumerate(shards):
+    for index, shard in enumerate(shards):
+        master_shard = shard['master']
+        slave_shards = shard.get('slaves', [])
+
         start, end = _get_logical_range(index, logical_count_per_shard)
         for logical_index in range(start, end):
             shard_name = _make_shard_name(shard_group=shard_group, shard_count=shard_count, shard_index=index, logical_index=logical_index)
-            replication_config = {
-                'master': _make_shard_database_url(
-                    origin=shard_config['master'], database_name=database_name, logical_index=logical_index, db_options=db_options
-                ),
-                'slaves': [
-                    _make_shard_database_url(
-                        origin=slave_database_url, database_name=database_name, logical_index=logical_index, db_options=db_options
-                    )
-                    for slave_database_url in shard_config.get('slaves', [])
-                ],
-                'conn_max_age': shard_options.get('conn_max_age', 0)
-            }
 
-            configuration.update(make_replication_configuration(key=shard_name, replication_config=replication_config, options={
-                'shard_group': shard_group, 'shard_number': logical_index
-            }))
+            _master_shard, _slave_shards = _make_shard_config(master_shard, slave_shards, database_name, logical_index, options)
+            configuration.update(make_replication_configuration(
+                key=shard_name, master=_master_shard, slaves=_slave_shards, conn_max_age=conn_max_age, shard_info={
+                    'shard_group': shard_group, 'shard_number': logical_index}
+            ))
 
     return configuration
 
 
-def make_replication_configuration(key: str, replication_config: Dict, options: Dict=None) -> Dict:
-    if replication_config.get('master', None) is None:
-        raise RequireMasterConfigException()
-
+def make_replication_configuration(key: str, master: Dict, slaves: List[Dict], conn_max_age: int, shard_info: Optional[Dict]=None) -> Dict:
     configuration = {}
-    conn_max_age = replication_config.get('conn_max_age', 0)
 
-    if options is None:
-        options = {}
+    if shard_info is None:
+        shard_info = {}
 
-    configuration[key] = _generate_database_config(database_url=replication_config['master'], conn_max_age=conn_max_age, **options)
-    for index, database_url in enumerate(replication_config.get('slaves', [])):
+    configuration[key] = _generate_database_config(
+        database_url=master['url'], conn_max_age=master.get('conn_max_age', conn_max_age), **shard_info
+    )
+
+    for index, slave in enumerate(slaves):
         slave_key = '%s_slave_%d' % (key, index)
-        configuration[slave_key] = _generate_database_config(database_url=database_url, conn_max_age=conn_max_age, is_replica_of=key)
+        configuration[slave_key] = _generate_database_config(
+            database_url=slave['url'], conn_max_age=slave.get('conn_max_age', conn_max_age), is_replica_of=key
+        )
 
     return configuration
+
+
+def _make_shard_config(master_shard: Dict, slave_shards: List, database_name: str, logical_index: int, options: Dict) -> Tuple[Dict, List]:
+    _master_shard = copy.deepcopy(master_shard)
+    _master_shard.update({
+        'url': _make_shard_database_url(
+            origin=master_shard['url'], database_name=database_name, logical_index=logical_index,
+            db_options={**options, **(_master_shard.get('options', {}))}
+        ),
+    })
+
+    _slave_shards = copy.deepcopy(slave_shards)
+    for slave_shard in _slave_shards:
+        slave_shard.update({
+            'url': _make_shard_database_url(
+                origin=slave_shard['url'], database_name=database_name, logical_index=logical_index,
+                db_options={**options, **(slave_shard.get('options', {}))}
+            )
+        })
+
+    return _master_shard, _slave_shards
 
 
 def _generate_database_config(